+++ source/menu/menu_code.cpp
@@ -14,7 +14,7 @@ void CMenu::_hideCode(bool instant)
 	m_btnMgr.hide(m_codeBtnAge, instant);
 	m_btnMgr.hide(m_codeLblTitle, instant);
 	for(u8 i = 0; i < ARRAY_SIZE(m_codeLblUser); ++i)
-		if(m_codeLblUser[i] != (u16)-1)
+		if(m_codeLblUser[i] != -1)
 			m_btnMgr.hide(m_codeLblUser[i], instant);
 	m_btnMgr.hide(m_codeLblAge, true);
 }
@@ -27,7 +27,7 @@ void CMenu::_showCode(void)
 	m_btnMgr.show(m_codeBtnBack);
 	m_btnMgr.show(m_codeLblTitle);
 	for(u8 i = 0; i < ARRAY_SIZE(m_codeLblUser); ++i)
-		if(m_codeLblUser[i] != (u16)-1)
+		if(m_codeLblUser[i] != -1)
 			m_btnMgr.show(m_codeLblUser[i]);
 	m_btnMgr.hide(m_codeLblAge, true);
 }
