+++ source/menu/menu_configsnd.cpp
@@ -24,7 +24,7 @@ void CMenu::_hideConfigSnd(bool instant)
 	m_btnMgr.hide(m_configSndBtnCFVolP, instant);
 	m_btnMgr.hide(m_configSndBtnCFVolM, instant);
 	for(u8 i = 0; i < ARRAY_SIZE(m_configSndLblUser); ++i)
-		if(m_configSndLblUser[i] != (u16)-1)
+		if(m_configSndLblUser[i] != -1)
 			m_btnMgr.hide(m_configSndLblUser[i], instant);
 }
 
@@ -49,7 +49,7 @@ void CMenu::_showConfigSnd(void)
 	m_btnMgr.show(m_configSndBtnCFVolP);
 	m_btnMgr.show(m_configSndBtnCFVolM);
 	for(u8 i = 0; i < ARRAY_SIZE(m_configSndLblUser); ++i)
-		if(m_configSndLblUser[i] != (u16)-1)
+		if(m_configSndLblUser[i] != -1)
 			m_btnMgr.show(m_configSndLblUser[i]);
 
 	m_btnMgr.setText(m_configSndLblGuiVolVal, wfmt(L"%i", m_cfg.getInt("GENERAL", "sound_volume_gui", 255)));
